Enhance thread logging in MatrixMultiplierImpl to include thread ID for better traceability

+++ server/src/server/MatrixMultiplierImpl.java
@@ -249,7 +249,7 @@ public class MatrixMultiplierImpl extends UnicastRemoteObject implements MatrixM
             if (rowEnd - rowStart <= threshold) {
                 int p = B[0].length, m = B.length;
                 // Registrar inicio del procesamiento de este bloque
-                logger.threadStart(rowStart, rowEnd, Thread.currentThread().getId()); // Cambiar para incluir threadId
+                logger.threadStart(rowStart, rowEnd, (int)Thread.currentThread().getId()); // Cambiar para incluir threadId
                 
                 for (int i = rowStart; i < rowEnd; i++) {
                     for (int j = 0; j < p; j++) {
@@ -263,7 +263,7 @@ public class MatrixMultiplierImpl extends UnicastRemoteObject implements MatrixM
                     logger.threadProgress(i, i - rowStart); // Cambiar para incluir currentRow
                 }
                 // Registrar finalización del bloque
-                logger.threadComplete(Thread.currentThread().getId()); // Cambiar para incluir threadId
+                logger.threadComplete((int)Thread.currentThread().getId()); // Cambiar para incluir threadId
                 // Actualizar progreso global
                 updateProgress(rowEnd - rowStart);
             } else {
